Add `patch` utils & rename `utils` to `paginate`
